DiseasePrediction/requirements.txt
numpy
pandas
fastapi
pydantic
uvicorn
rich
python-dotenv
scikit-learn
matplotlib
xgboost
seaborn
imblearn
joblib